CB-16926 Redbeams provisioning flow restart causes duplicate deployment submission on Azure
In redbeams, when the provisioning flow restarts, the database template deployment is sometimes posted twice, causing an Azure error.

This commit has a generalized solution for the problem. If a template needs to be submitted, then the code
- checks if there is a template already in the given resource group (RG) with the given name. If no, then a new teplate is submitted
- if the template exists but is terminal (succeeded, failed, recently deleted etc), then a new template is submitted. If the previous and current templates had to provisioned the same resources, no change is done, if the resources still exist.
- if the template is in progress, then a poller is started.

The poller in this case is a new databasePoller.It does not check the existence of the database as after the database provisioning either service or private endpoints are also created.

The new solution was added only for database provisioning. If needed it can be added to other places in the code.